Auto-generate lookup tables for unifont.c and dejavu.c. The generated lookup tables are included by these C files and pointed-to by new Font.code_to_id which has type const __flash uint8_t*. The auto-generated files unifont-lookup.h and dejavu-lookup.h are generated on the fly by new Python script gen-lookup.py, and for now these files are not in the Git repo. This means that the user must have a Python interpreter at hand.
font.c::getGlyphAddress() uses the lookup table to speed to glyph lookup,
provided font->code_to_id is availabvle -- which is always the case for now.
This means that the linear lookup in getGlyphAddress() is actually dead code.

The type of Font.code has been changed from uint16_t to code_t, which
makes it easier to switch char codes from a 16-bit type to an 8-bit type.
code_t is typedef'ed to uint16_t.

Code sizes with avr-gcc v8 are:
* 16956 bytes with code_t = uint16_t
* 16716 bytes with code_t = uint8_t
which is a difference of 240 bytes.  theUnifont contains 190 gyphs, and
theDejaVu contains 28 glyphs, hence the expected gain of switching
code_t from uint16_t to uint8_t is 218 bytes at leat.

thermidity.elf depends on all headers, so respective dependencies were added
in the Makefile.

thermidity-avr/
	* gen-lookup.py: New Python script.

	* Makefile ($(TARGET).elf): Depend on: Makefile, bitmaps.h, dejavu.h,
	display.h, eink.h, font.h, meter.h, pins.h, spi.h, sram.h, unifont.h,
	usart.h, utils.h, unifont-lookup.h, dejavu-lookup.h.
	(unifont-lookup.h): New rule to auto-generate it from unifont.c.
	(dejavu-lookup.h): New rule to auto-generate it from dejavu.c.
	(clean): Add unifont-lookup.h dejavu-lookup.h.

	* font.h (code_t): New typedef to uint16_t.
	(Font) [code]: Change element type to code_t.
	[code_to_id]: New element pointing to uint8_t[256] lookup, or NULL.
	(getGlyphAddress) [code]: Change param type to code_t.

	* font.c (getGlyphAddress): Use lookup table font->code_to_id if
	available.
	[code]: Change param type to code_t.

	* dejavu.c (dejavu-lookup.h): New include.
	(theDejaVu): Add 4th element lookup_dejavu to initilizer.

	* unifont.c (unifont-lookup.h): New include.
	(theUnifont): Add 4th element lookup_unifont to initilizer.

	* display.c (writeGlyph) [code]: Change param type to code_t.
	(writeString): Pass code as code_t to writeGlyph.

	* bitmaps.c: Clean up garbage comment.